requirements.txt
beautifulsoup4
selenium
python-dotenv
requests
webdriver-manager
